feat: generate field cells and flags on first click
