+++ src/style.scss
@@ -223,6 +223,7 @@ body::-webkit-scrollbar-thumb {
   }
 
   &_field {
+    aspect-ratio: 1 / 1;
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
@@ -230,10 +231,15 @@ body::-webkit-scrollbar-thumb {
     border: 5px solid salmon;
     margin-top: 20px;
     background-color: rgba(168, 221, 89, 0.8);
+    font-size: 4rem;
 
     .active {
       background-color: rgba(89, 221, 170, 0.8);
       // background-color: rgba(89, 142, 221, 0.8);
+    }
+
+    &:hover {
+      cursor: pointer;
     }
   }
